css/css-page/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Paged Media properties</title>
<link rel="help" href="https://drafts.csswg.org/css-page-3/#property-index">
<meta name="assert" content="Properties should not inherit.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="reference"></div>
<div id="container">
  <div id="target"></div>
</div>
<script>
assert_not_inherited('page', 'auto', 'PaGeTyPeNaMe');
</script>
</body>
</html>
