LayoutTests/imported/w3c/web-platform-tests/css/css-page/inheritance-expected.txt


FAIL Property page has initial value auto assert_equals: expected "auto" but got ""
FAIL Property page does not inherit assert_equals: expected "koko" but got ""
